flight: Move TestFlightServer to its own .h/.cc pair
